pyproject.toml
[tool.poetry]
name = "business-dashboard-gcp-app"
version = "0.1.0"
description = ""
authors = ["miroslaw-steblik <miroslawsteblik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
faker = "^25.0.1"
pandas = "^2.2.2"
streamlit = "^1.34.0"
plotly = "^5.22.0"
google-cloud-bigquery = "^3.23.1"
db-dtypes = "^1.2.0"
requests = "^2.32.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
